+++ client/App.js
@@ -1,20 +1,18 @@
 
 import React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
+import { MainContainer } from './Components/MainContainer/MainContainer';
+
+/* import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { Landing } from './Components/Landing/Landing';
 import { Home } from './Components/Home/Home';
+import { LookProducts } from './Components/LookProducts/LookProducts'; */
+
 
 const App= () => {
-
-const Stack = createNativeStackNavigator();
+/* 
+const Stack = createNativeStackNavigator(); */
   return (
-    <NavigationContainer>
-      <Stack.Navigator initialRouteName='Landing' screenOptions={{ headerShown: false }} >
-        <Stack.Screen name='Landing' component={Landing} />
-        <Stack.Screen name='Home' component={Home}/>
-      </Stack.Navigator>
-    </NavigationContainer>
+    <MainContainer/>
   );
 };
