+++ client/Components/Cart/Cart.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { View, Text } from 'react-native'
+
+export const Cart = () => {
+  return (
+    <View>
+        <Text>Carrito de Compras</Text>
+    </View>
+  )
+}
+
